# src/frontend/shared/openapi/wnav-openapi.yaml
openapi: "3.1.0"
info:
  title: WNAV API
  version: 1.0.0
  description: |
    作業ナビゲーションシステム REST API。terminal-api（ポート 8080）と master-api（ポート 8081）の統合仕様。
    エラーは RFC 9457 Problem Details 形式で返却する。
  contact:
    name: WNAV Engineering
    email: engineering@wnav.example.com
servers:
  - url: http://localhost:8080/api/v1
    description: terminal-api ローカル開発
  - url: http://localhost:8081/api/v1
    description: master-api ローカル開発
  - url: https://wnav-terminal.factory.local/api/v1
    description: terminal-api 本番
  - url: https://wnav-master.factory.local/api/v1
    description: master-api 本番
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      operationId: postAuthLogin
      summary: ログイン（API-auth-001）
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthLoginRequest' }
      responses:
        '200': { description: 認証成功, content: { application/json: { schema: { $ref: '#/components/schemas/AuthLoginResponseEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '423': { $ref: '#/components/responses/Locked' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: postAuthRefresh
      summary: トークン更新（API-auth-002）
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRefreshRequest' }
      responses:
        '200': { description: 更新成功, content: { application/json: { schema: { $ref: '#/components/schemas/AuthRefreshResponseEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      operationId: postAuthLogout
      summary: ログアウト（API-auth-003）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '204': { description: ログアウト完了 }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /electronic-signs:
    post:
      tags: [electronic-signs]
      operationId: postElectronicSign
      summary: 電子サイン作成（API-electronic-signs-001）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateElectronicSignRequest' }
      responses:
        '201': { description: 作成成功, content: { application/json: { schema: { $ref: '#/components/schemas/ElectronicSignEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
    get:
      tags: [electronic-signs]
      operationId: listElectronicSigns
      summary: 電子サイン一覧（API-electronic-signs-003）
      parameters:
        - { name: work_execution_id, in: query, schema: { type: string } }
        - { name: signer_id, in: query, schema: { type: string } }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
      responses:
        '200': { description: 一覧, content: { application/json: { schema: { $ref: '#/components/schemas/ElectronicSignListEnvelope' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 で署名された JWT を Bearer トークンとして送信する
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid, description: UUID v7 形式 }
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
  responses:
    Unauthorized:
      description: 認証失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    ValidationError:
      description: バリデーション失敗
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Locked:
      description: アカウントロック
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
...
    ResponseMeta:
      type: object
      required: [request_id, server_time, api_version]
      properties:
        request_id: { type: string, format: uuid }
        server_time: { type: string, format: date-time }
        api_version: { type: string, enum: [v1] }
    PageMeta:
      type: object
      required: [total, page, per_page, total_pages]
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        total_pages: { type: integer }
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        error_id: { type: string }
        violations:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
              value: {}
...
    AuthLoginRequest:
      type: object
      required: [login_id, password, device_id, factory_id]
      properties:
        login_id: { type: string }
        password: { type: string }
        device_id: { type: string, format: uuid }
        factory_id: { type: string, format: uuid }
    AuthLoginResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, refresh_expires_in, roles, user_id, factory_id]
      properties:
        access_token: { type: string }
        refresh_token: { type: string, format: uuid }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer }
        refresh_expires_in: { type: integer }
        roles: { type: array, items: { type: string } }
        user_id: { type: string, format: uuid }
        factory_id: { type: string, format: uuid }
    AuthLoginResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AuthLoginResponse' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    AuthRefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: { type: string, format: uuid }
    AuthRefreshResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token: { type: string }
        token_type: { type: string }
        expires_in: { type: integer }
    AuthRefreshResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/AuthRefreshResponse' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...
    CreateElectronicSignRequest:
      type: object
      required: [signer_id, signed_content_hash, pin_hash, context_type, context_id, timestamp_client, device_signature]
      properties:
        signer_id: { type: string, format: uuid }
        signed_content_hash: { type: string }
        pin_hash: { type: string }
        context_type: { type: string, enum: [step_sign, work_complete_sign, approval_sign, quality_check_sign] }
        context_id: { type: string, format: uuid }
        step_id: { type: string, format: uuid }
        timestamp_client: { type: string, format: date-time }
        device_signature: { type: string }
    ElectronicSign:
      type: object
      properties:
        sign_id: { type: string, format: uuid }
        signer_id: { type: string, format: uuid }
        signed_content_hash: { type: string }
        context_type: { type: string }
        context_id: { type: string, format: uuid }
        signed_at: { type: string, format: date-time }
        hash_chain_block_id: { type: string, format: uuid }
        hash_chain_value: { type: string }
    ElectronicSignEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/ElectronicSign' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    ElectronicSignListEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/ElectronicSign' } }
        meta:
          allOf:
            - $ref: '#/components/schemas/ResponseMeta'
            - type: object
              properties:
                pagination: { $ref: '#/components/schemas/PageMeta' }
...